requirements.txt
# MetaTrader5  # Install manually if needed
pandas>=1.5.0
numpy>=1.21.0
ta>=0.10.2
scikit-learn>=1.1.0
xgboost>=1.6.0
scipy>=1.9.0
beautifulsoup4>=4.11.0
requests>=2.28.0
python-dotenv>=0.19.0
pytz>=2022.1
joblib>=1.1.0